cluster runner now support adding slaves

+++ redisson/src/test/java/org/redisson/RedisRunner.java
@@ -191,6 +191,7 @@ public class RedisRunner {
     protected static RedisRunner.RedisProcess defaultRedisInstance;
     private static int defaultRedisInstanceExitCode;
 
+    private String path = "";
     private String defaultDir = Paths.get("").toString();
     private boolean nosave = false;
     private boolean randomDir = false;
@@ -459,6 +460,7 @@ public class RedisRunner {
     public RedisRunner dir(String dir) {
         if (!randomDir) {
             addConfigOption(REDIS_OPTIONS.DIR, dir);
+            this.path = dir;
         }
         return this;
     }
@@ -824,6 +826,10 @@ public class RedisRunner {
     public String defaultDir() {
         return this.defaultDir;
     }
+    
+    public String dir() {
+        return this.path;
+    }
 
     public String getInitialBindAddr() {
         return bindAddr.size() > 0 ? bindAddr.get(0) : "localhost";
@@ -849,7 +855,7 @@ public class RedisRunner {
 
     public boolean deleteClusterFile() {
         File f = new File(clusterFile);
-        if (f.exists()) {
+        if (f.exists() && isRandomDir()) {
             System.out.println("REDIS RUNNER: Deleting cluster config file " + f.getAbsolutePath());
             return f.delete();
         }
